fix(analyze): use u64 accumulators in tier2 chroma sharpness for 32-bit safety
`image_sharpness_breakdown` accumulates per-pixel `gradient_diff_ycbcr`
results into `sumh`/`sumv` and at fragment boundaries divides by
`fragment_height * width`. Per-pixel values can reach ~16M (squared
chroma 2nd-difference times the boost factor) and a single fragment
spans up to 16 rows × image width × 31 horizontal positions, so the
sum can grow past ~7.9 × 10⁹ on a 64×64 stripe pattern — overflowing
`usize = u32` on 32-bit targets.

Fixes the `analyze::tests::vstripes_have_high_horiz_chroma_zero_vert`
panic ("attempt to add with overflow") in i686 CI. Pre-existing on
main; surfaced because PR #115 dropped the i686 ultrahdr-feature
disable that masked unrelated failures.

+++ zenjpeg/src/analyze/tier2_chroma.rs
@@ -98,8 +98,11 @@ fn image_sharpness_breakdown(
     stream.fetch_into(1, &mut row1);
     stream.fetch_into(2, &mut row2);
 
-    let mut sumh: (usize, usize) = (0, 0);
-    let mut sumv: (usize, usize) = (0, 0);
+    // u64 (not usize) so 32-bit builds don't overflow:
+    // gradient_diff_ycbcr can return values up to ~437K per pixel, and
+    // fragments span thousands of pixels — summing in usize=u32 wraps.
+    let mut sumh: (u64, u64) = (0, 0);
+    let mut sumv: (u64, u64) = (0, 0);
     let mut max_sumh: (u32, u32) = (0, 0);
     let mut max_sumv: (u32, u32) = (0, 0);
     let mut max_diff: (u32, u32) = (0, 0);
@@ -140,15 +143,15 @@ fn image_sharpness_breakdown(
                 max_diff.1 = h.1;
             }
 
-            sumh.0 += h.0 as usize;
-            sumh.1 += h.1 as usize;
-            sumv.0 += v.0 as usize;
-            sumv.1 += v.1 as usize;
+            sumh.0 += h.0 as u64;
+            sumh.1 += h.1 as u64;
+            sumv.0 += v.0 as u64;
+            sumv.1 += v.1 as u64;
         }
 
         fragment_height += 1;
         if fragment_height >= fragment_max_height {
-            let denom = fragment_height * width;
+            let denom = (fragment_height * width) as u64;
             max_sumh.0 = max_sumh.0.max((sumh.0 / denom) as u32);
             max_sumh.1 = max_sumh.1.max((sumh.1 / denom) as u32);
             max_sumv.0 = max_sumv.0.max((sumv.0 / denom) as u32);
@@ -170,7 +173,7 @@ fn image_sharpness_breakdown(
         stream.fetch_into(need_y2 as u32, &mut row2);
     }
     if fragment_height > 16 {
-        let denom = fragment_height * width;
+        let denom = (fragment_height * width) as u64;
         max_sumh.0 = max_sumh.0.max((sumh.0 / denom) as u32);
         max_sumh.1 = max_sumh.1.max((sumh.1 / denom) as u32);
         max_sumv.0 = max_sumv.0.max((sumv.0 / denom) as u32);
